modularize output table writing code
